+++ themes/HeroesLounge-Theme/layouts/cleanNoTheme.htm
@@ -6,8 +6,6 @@ description = "Clean No Nav"
 
 [googleTracker]
 
-[SideNav]
-
 [session]
 security = "all"
 ==
