+++ themes/HeroesLounge-Theme/layouts/plain-with-sidebar.htm
@@ -6,8 +6,6 @@ description = "Plain layout for sections"
 
 [googleTracker]
 
-[SideNav]
-
 [SetTimezone]
 
 [session]
@@ -18,9 +16,6 @@ security = "all"
 <body class="{{ this.page.id }} page page-{{ this.page.id }} layout-{{ this.layout.id }} {% placeholder bodyclass title='Body class' %}">
 
     {% partial 'base/head' %} 
-    <div class="sticky-top float-left" style="top:90px">
-        {% component 'SideNav' %}
-    </div>
     <div class="section page-content bg-white">
         <div class="container">
             <div class="row">
